requirements.txt
Flask==2.0.1
pytest==6.2.5
rq==1.10.1
Pillow==9.0.1
gunicorn==20.1.0